Renamed androidPredictiveBackAnimatable to androidPredictiveBackAnimatableV1
